+++ Xlab.Test.Api/Businesses/ResourceRepresentations/BusinessStarsResourceRepresentation.cs
@@ -2,12 +2,14 @@ using Xlab.Test.Domain.Businesses;
 
 namespace Xlab.Test.Api.Businesses;
 
+// star values set to double as mongodb driver was freaking out about them when set to int like they should be and I'm not sure why
+
 public class BusinessStarsResourceRepresentation
 {
-    public int Beer { get; set; }
-    public int Atmosphere { get; set; }
-    public int Amenities { get; set; }
-    public int Value { get; set; }
+    public double Beer { get; set; }
+    public double Atmosphere { get; set; }
+    public double Amenities { get; set; }
+    public double Value { get; set; }
 
     public static BusinessStarsResourceRepresentation From(Business business) => new()
     {
