+++ Xlab.Test.Domain/Businesses/Business.cs
@@ -1,5 +1,7 @@
 ﻿namespace Xlab.Test.Domain.Businesses;
 
+// star values set to double as mongodb driver was freaking out about them when set to int like they should be and I'm not sure why
+
 public record Business(
     Guid id,
     string name,
@@ -12,10 +14,10 @@ public record Business(
     string address,
     string phone,
     string twitter,
-    int stars_beer,
-    int stars_atmosphere,
-    int stars_amenities,
-    int stars_value,
+    double stars_beer,
+    double stars_atmosphere,
+    double stars_amenities,
+    double stars_value,
     DateTime date,
     IEnumerable<string> tags
 );
